MLPDR-main/requirements.txt
arabic-reshaper==2.1.3
cycler==0.10.0
decorator==4.4.2
future==0.18.2
imageio==2.9.0
imutils==0.5.4
kiwisolver==1.3.1
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
opencv-contrib-python==4.5.1.48
opencv-python==4.5.2.52
Pillow==8.2.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pytesseract==0.3.7
python-bidi==0.4.2
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.3
six==1.15.0
tifffile==2021.4.8
Flask
